Moved IPlayer, Dealer, and Player to their own files
