+++ blackjack/blackjack/IPlayer.cs
@@ -1,0 +1,8 @@
+namespace blackjack
+{
+    internal interface IPlayer
+    {
+        public void Hit();
+        public void Stay();
+    }
+}
